Some first ideas to have semi-automatic masters and slaves
Seems to be a dead end. If a host is master or slave or forward for this domain is actually encoded in the domain data bag itself. If its in the nameservers list, its either a slave or a master, if its also in the allow_transfer list, its a slave. Otherwise its a forward server.